==> Final Notebooks/00 download_dataset.ipynb <==
! mkdir dataset

# Download main set

! curl -o dataset/spotify_million_playlist_dataset.zip 'https://spotify-million-playlist-dataset.s3.eu-central-1.wasabisys.com/files/spotify_million_playlist_dataset.zip?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=N5Y368R6EBCIGK4YBS1R%2F20210520%2Feu-central-1%2Fs3%2Faws4_request&X-Amz-Date=20210520T091621Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=4fc13f23c610aeb4a5ad740ecd0d2745730cf690303fdfe2e9ced6c476d4f4e9'

# Download challenge set

! curl -o dataset/spotify_million_playlist_dataset_challenge.zip 'https://spotify-million-playlist-dataset.s3.eu-central-1.wasabisys.com/files/spotify_million_playlist_dataset_challenge.zip?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=N5Y368R6EBCIGK4YBS1R%2F20210520%2Feu-central-1%2Fs3%2Faws4_request&X-Amz-Date=20210520T091621Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=d59452902123500acdccc057f8bf65649c52bc92d07276412ebe38b1e544f2be'

# Unzip files

! unzip -q dataset/spotify_million_playlist_dataset.zip -d dataset/spotify_million_playlist_dataset
! unzip -q dataset/spotify_million_playlist_dataset_challenge.zip -d dataset/spotify_million_playlist_dataset_challenge

# Copy files to cloud storage

! gsutil cp dataset/spotify_million_playlist_dataset/data/* gs://thesis_apc_bucket/data/mpd
! gsutil cp dataset/spotify_million_playlist_dataset_challenge/challenge_set.json gs://thesis_apc_bucket/data/challenge_set.json
