==> Final Notebooks/07 co-occurrence_features.ipynb <==
import pyspark
import pandas as pd
import numpy as np
import scipy.sparse as sp

# Initialize pyspark

from pyspark.sql import SparkSession

spark = SparkSession.builder \
                    .appName("LightFM Features") \
                    .getOrCreate()

# Load train dataframe

path_to_df = 'gs://thesis_apc_bucket/df_data2'

df_data = spark.read.orc(path_to_df + '/df_train.orc')

# Drop duplicate rows in pid and tid columns

df_data = df_data.dropDuplicates(['pid', 'tid'])

# Get number of playlists and number of tracks

tid_max = df_data.agg({'tid': 'max'}).collect()[0]['max(tid)']

num_tracks = tid_max + 1

num_tracks

data = df_data.toPandas()

%%time

# Calculate occurrences

from collections import defaultdict
from pyspark.sql.functions import col

co_occurrence_2 = [defaultdict(int) for i in range(num_tracks)]
# co_occurrence_3 = [defaultdict(lambda: defaultdict(int)) for i in range(num_tracks)]
occurrence = [0 for i in range(num_tracks)]
# pids = data.select('pid').distinct().rdd.map(lambda x: x[0]).collect()

for q, (_, df) in enumerate(data.groupby('pid')):
    if q % 10000 == 0:
        print(q / 10000)
    tids = list(df.tid)
    for i in tids:
        occurrence[i] += 1
    for k, i in enumerate(tids):
        for j in tids[k + 1:]:
#             temp_tids = [i] + [j]
#             temp_tids.sort()
            co_occurrence_2[i][j] += 1
            co_occurrence_2[j][i] += 1
#             for l in tids[k + 2:]:
#                 temp_tids += [l]
#                 temp_tids.sort()
#                 co_occurrence_3[temp_tids[0]][temp_tids[1]][temp_tids[2]] += 1

# Group pids by tid

data_tracks = data.groupby('tid').pid.apply(list)

# Define function to get co-occurrence-3 values from data

def find_co_occurrence_3(i, j, k):
    set_i = set(data_tracks[i])
    set_j = set(data_tracks[j])
    set_k = set(data_tracks[k])
    return len(set.intersection(set_i, set_j, set_k))

def get_f(i, f):
    if len(i) == 0:
        return -1
    else:
        return f(i)
    
def create_co_occurrence_features(df):
    
    df = df.toPandas()
    
    pids = df.pid.unique()
    seed = data[data.pid.isin(pids)]
    tid_seed = seed.groupby('pid').tid.apply(list)
    
    co_occurrence_seq = []
    print('Creating features...')
    count = 0
    for pid, tid in df[['pid', 'tid']].values:
        tid = int(tid)
        count += 1
        if (count % 1000000 == 0):
            print(count / 1000000)
        tracks = tid_seed.get(pid, [])
        co_occurrence_seq.append(np.array([co_occurrence_2[tid][i] for i in tracks]))
        
    df['co_occurrence_max'] = [get_f(i, np.max) for i in co_occurrence_seq]
    df['co_occurrence_min'] = [get_f(i, np.min) for i in co_occurrence_seq]
    df['co_occurrence_mean'] = [get_f(i, np.mean) for i in co_occurrence_seq]
    df['co_occurrence_median'] = [get_f(i, np.median) for i in co_occurrence_seq]
    
    co_occurrence_seq = []
    print('Creating normalized features...')
    count = 0
    for pid, tid in df[['pid', 'tid']].values:
        tid = int(tid)
        count += 1
        if (count % 1000000 == 0):
            print(count / 1000000)
        tracks = tid_seed.get(pid, [])
        seq = np.empty(0)   # Initiate array
        for k in range(len(tracks) - 1):
            seq = np.append(seq, find_co_occurrence_3(tid, tracks[k], tracks[k+1]) / co_occurrence_2[tid][tracks[k]] 
                            if co_occurrence_2[tid][tracks[k]] > 0 else 0)
        co_occurrence_seq.append(seq)
#         if (count % 1000000 == 0):
#             print(co_occurence_seq)
        
    df['co_occurrence_norm_max'] = [get_f(i, np.max) for i in co_occurrence_seq]
    df['co_occurrence_norm_min'] = [get_f(i, np.min) for i in co_occurrence_seq]
    df['co_occurrence_norm_mean'] = [get_f(i, np.mean) for i in co_occurrence_seq]
    df['co_occurrence_norm_median'] = [get_f(i, np.median) for i in co_occurrence_seq]
    
    res = spark.createDataFrame(df)
    return res

# Load candidates

import pandas as pd

path_to_df = 'gs://thesis_apc_bucket/df_data2'

df_train = spark.read.orc(path_to_df + '/ii_candidate.orc')
df_val = spark.read.orc(path_to_df + '/iii_candidate.orc')
df_test = spark.read.orc(path_to_df + '/test_candidate.orc') 

%%time

df_train = create_co_occurrence_features(df_train)

%%time

df_train.describe().show()

df_train.write.mode('overwrite').orc(path_to_df + '/ii_co_occurrence_features')

%%time

df_val = create_co_occurrence_features(df_val)

%%time

df_val.show()

df_val.write.mode('overwrite').orc(path_to_df + '/iii_co_occurrence_features')

%%time

df_test = create_co_occurrence_features(df_test)

%%time

df_test.show()

%%time

df_train.write.mode('overwrite').orc(path_to_df + '/ii_co_occurrence_features')
df_val.write.mode('overwrite').orc(path_to_df + '/iii_co_occurrence_features')
df_test.write.mode('overwrite').orc(path_to_df + '/test_co_occurrence_features')
