==> Final Notebooks/09 xgboost_experiment.ipynb <==
import os
os.environ['KMP_DUPLICATE_LIB_OK']='True'

import xgboost
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plot

# Load dataframes

train = pd.read_hdf('df_data2/train_final.hdf')
val = pd.read_hdf('df_data2/val_final.hdf')
test = pd.read_hdf('df_data2/test_final.hdf')

xgtrain = xgboost.DMatrix(train.drop(['pid', 'tid', 'target'], axis=1), train.target)

xgval = xgboost.DMatrix(val.drop(['pid', 'tid', 'target'], axis=1), val.target)

xgtest = xgboost.DMatrix(test.drop(['pid', 'tid'], axis=1))

%%time

# Train XGBoost model
# Manual parameter tuning

max_auc = 0.0
best_params = None

# Tuning max-depth
for max_depth in [5, 6, 7, 8, 9]:
    print('max_depth: {}'.format(max_depth))
    
    result = {}
    
    params = {
        'objective':'binary:logistic',
        'eta':0.1,
        'booster':'gbtree',
        'max_depth': max_depth,
        'nthread':16,
        'seed':1,
        'eval_metric':'auc',
    }

    model = xgboost.train(
        params=list(params.items()),
        early_stopping_rounds=30, 
        verbose_eval=10, 
        dtrain=xgtrain,
        evals=[(xgtrain, 'train'), (xgval, 'test')],
        evals_result=result,
        num_boost_round=300,
    )
    best_score = float(model.attributes()['best_score'])
    last_train = result['train']['auc'][-1]
    last_test = result['test']['auc'][-1]
    print('Last train-AUC: {}, last test-AUC: {}, best iteration: {}, best score: {}'\
          .format(last_train, last_test, model.attributes()['best_iteration'], model.attributes()['best_score']))
    if best_score > max_auc:
        max_auc = best_score
        best_params = max_depth
        model.save_model('models2/trial2_best.model')
        
print('Best params: max_depth={}'.format(best_params))

# %%time

# # Grid search using xgboost.cv

# best_params = None

# grid_params = [
#     (max_depth, min_child_weight)
#     for max_depth in [7, 8, 9]
#     for min_child_weight in [0.5, 1, 1.5, 2]
# ]

# max_auc = 0.0

# for max_depth, min_child_weight in grid_params:
#     result = {}
    
#     print("CV with max_depth={}, min_child_weight={}".format(
#                              max_depth,
#                              min_child_weight))
#     # Update our parameters
#     params = {
#         'objective':'binary:logistic',
#         'eta':0.1,
#         'booster':'gbtree',
#         'max_depth': max_depth,
#         'min_child_weight': min_child_weight,
#         'nthread':16,
#         'seed':1,
#         'eval_metric':'auc',
#     }
# #     params['max_depth'] = max_depth
# #     params['min_child_weight'] = min_child_weight
#     print('Run CV...')
#     # Run CV
#     model = xgboost.train(
#         params=list(params.items()),
#         dtrain=xgtrain,
#         verbose_eval=10, 
#         num_boost_round=300,
#         evals_result=result,
#         evals=[(xgtrain, 'train'), (xgval, 'test')],
#         early_stopping_rounds=30,
#     )
#     # Update best AUC
#     best_score = float(model.attributes()['best_score'])
# #     mean_auc = cv_results['test-auc-mean'].max()
# #     boost_rounds = cv_results['test-auc-mean'].argmax()
# #     print("\tAUC {} for {} rounds".format(mean_auc, boost_rounds))
#     last_train = result['train']['auc'][-1]
#     last_test = result['test']['auc'][-1]
#     print('Last train-AUC: {}, last test-AUC: {}, best iteration: {}, best score: {}'\
#           .format(last_train, last_test, model.attributes()['best_iteration'], model.attributes()['best_score']))
#     if best_score > max_auc:
#         max_auc = best_score
#         best_params = (max_depth, min_child_weight)
#         model.save_model('models2/trial3_best.model')
        
# print("Best params: {}, {}, AUC: {}".format(best_params[0], best_params[1], max_auc))

# Use trial 3's best model to generate candidates

bst = xgboost.Booster({'nthread': 16})  # init model
bst.load_model('models2/trial2_best.model')  # load data

# Initialize pyspark

from pyspark.sql import SparkSession

spark = SparkSession.builder \
                    .master('local[*]')\
                    .appName("LightFM Features") \
                    .getOrCreate()

path_to_df = 'gs://thesis_apc_bucket/df_data2'

df_tracks = spark.read.orc(path_to_df + '/df_tracks.orc')

test['score'] = bst.predict(xgtest)
test = test.sort_values(['pid', 'score'], ascending=[True, False])
recs = test.groupby('pid').tid.apply(lambda x: x.values[:500])
tracks_info = df_tracks.toPandas()
track_uri = tracks_info.track_uri

submission = open('submissions/submission-ps2.csv', 'w')
submission.write('team_info,main,Exin,martin.alinggajaya@gmail.com\n')

for pid, tids in recs.items():
    submission.write('{}, '.format(pid) + ', '.join(track_uri.loc[tids].values) + '\n')
    
submission.close()

# Show best features

feature_gain = bst.get_score(importance_type='gain')
sorted_feature_gain = {k: v for k, v in sorted(feature_gain.items(), key=lambda x: x[1], reverse=True)}
sorted_feature_gain
